Migrate parsing of WHERE clauses to its own module
This is in anticipation of extending Filter functionality, which
will require importing functions from 'common', which would create
circular imports, because parse_where uses classes from 'models'.
This logic will eventually be implemented in classes from 'models',
so this refactoring is temporary until the migration to the new
structure is complete.
